requirements.txt
fastapi
uvicorn
python-dotenv
playwright
playwright_stealth
setuptools
aiohttp
supabase

